requirements.txt
tkinter
pillow
plyer
tkcalendar
requests 